requirements_dev.txt
pip
wheel
flake8
tox
coverage
Sphinx
twine
black
pylint
pylint-flask
pylint-flask-sqlalchemy
mypy
doc8
rstcheck

Click
pytest
pytest-cov
pytest-runner
